[Chore #161203421] restructure models separate dir
